Axum project started
Started an Axum app that will be an optimization over the spring application, for now is not the main application, but it intends to be.

This application as the cors set up for the methods get, post, put, patch and delete, allows for any origin and header (for now).

It also implements sea_orm, an orm for rust that allows for async opperations.